Rename package; re-autospec
'poppler-data' is also a subpackage of 'poppler'. This situation is
problematic for tools that consume clr-bundle definitions (like mixer),
since it is ambiguous whether 'poppler-data' refers to (a) the
'poppler-data' package or (b) the subpackage of 'poppler'.

To resolve, start a new naming convention: add a "-clr-rename" suffix to
the package name whenever the upstream project name has a suffix that
autospec will potentially use for a package name that lacks that suffix.

So, for this case, rename 'poppler-data' to 'poppler-data-clr-rename'.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>